move surface class to surface.py
